Initial setup of the vllm source
Cherry-picked from 98bbb835 on feature/vllm-integration. Resolved against
dev-4.x's Phase A type split:
- bridge.py: kept the shim (TransformerBridge content lives in
  transformer_bridge.py now).
- sources/transformers.py: accepted the scaffold's package restructure
  (transformers.py → transformers/{__init__,helpers,source}.py +
  _hf_format.py).
- sources/transformers/source.py: re-applied Phase A's explicit
  TransformersDriver construction in boot().

sources/vllm/source.py's boot_vllm currently sets bridge._forward_impl,
which Phase A removed. The function is dead until the next chunk lifts
it onto VLLMDriver + RemoteBridge — landing as-is so the scaffold tree is
on dev-4.x for incremental Phase B work.
