+++ transformer_lens/model_bridge/sources/__init__.py
@@ -8,9 +8,11 @@ from transformer_lens.model_bridge.sources.transformers import (
     check_model_support,
     list_supported_models,
 )
+from transformer_lens.model_bridge.sources.vllm import boot_vllm
 
 __all__ = [
     "boot",
+    "boot_vllm",
     "list_supported_models",
     "check_model_support",
 ]
